+++ tests/test_stats_manager.py
@@ -141,7 +141,7 @@ def test_detector_metrics(test_video_file):
         duration = FrameTimecode('00:00:20', video_fps)
 
         video_manager.set_duration(start_time=start_time, end_time=duration)
-        video_manager.set_downscale_factor()
+        scene_manager.set_auto_downscale()
         video_manager.start()
         scene_manager.detect_scenes(frame_source=video_manager)
 
@@ -260,7 +260,7 @@ def test_save_load_from_video(test_video_file):
         duration = FrameTimecode('00:00:20', video_fps)
 
         video_manager.set_duration(start_time=start_time, end_time=duration)
-        video_manager.set_downscale_factor()
+        scene_manager.set_auto_downscale()
         video_manager.start()
         scene_manager.detect_scenes(frame_source=video_manager)